==> Final_Project/02.Get_Transfermarkt_Historicals.ipynb <==
from __future__ import print_function, division
import sys
from bs4 import BeautifulSoup
import requests

requests.__path__

from selenium import webdriver
from selenium.webdriver.common.keys import Keys
import time
import os
import random

import pandas as pd
import pickle

chromedriver = "/Users/georgevarelas/Downloads/chromedriver" # path to the chromedriver executable
os.environ["webdriver.chrome.driver"] = chromedriver

def get_historical_prices(link):
    url = "https://www.transfermarkt.com" + link.replace('profil', 'marktwertverlauf')
    driver = webdriver.Chrome(chromedriver)
    driver.get(url)
    
    time.sleep(2)
    soup = BeautifulSoup(driver.page_source, 'html.parser')
    
    script = soup.find_all('script')
    
    data = (str(script[len(script)-1]))
    data = data.split("'series':[{")[1].split("\'y")
    
    age = []
    value = []
    club = []
    date = []

    for i in range(1, len(data)):
        value.append(data[i].split(':')[1].split(',')[0])
        age.append(data[i].split(':')[3].split(',')[0])
        club.append(data[i].split(':')[2].split(',')[0].replace("'", "").replace("\\x20", " ").replace("\\x2D", " "))

        day = data[i].split(':')[5].split(',')[:-1]
        date.append(day[0][1:4] + " " + day[0][-2:] + ", "+ day[1][-5:-1])
    
    

    
    find_agents = soup.find_all("div", {"class":"row"})
    try:
        agent = str(find_agents).split('<span class="dataItem">Agent:</span>')[1]
        agent_name = agent.split('title="')[1].split('"')[0]
    except IndexError:
        agent_name = 'NAN'
    
    
    
    historicals = list(zip(*[date, value, age, club])) + [agent_name]
    
    driver.close()
    
    return historicals

get_historical_prices("/raheem-sterling/profil/spieler/134425")

players = pd.read_pickle('/Users/georgevarelas/Luther/Players_Transfermarkt.pkl')

players.head(3)

d = {}

not_appended = []

for i in (players['Profile_Link']):
    count = 0
    if count % 25 == 0:
        f = open("players_dictionary.pkl","wb")
        pickle.dump(d,f)
        f.close()
    try:
        d[i] = get_historical_prices(i)
        count += 1
    except IndexError:
        print(i)
        not_appended.append(i)
        continue
        
print(len(not_appended))

second_attempt_failed = []

for i in not_appended:
    count = 0
    if count % 25 == 0:
        f = open("players_dictionary.pkl","wb")
        pickle.dump(d,f)
        f.close()
    try:
        d[i] = get_historical_prices(i)
        count += 1
    except IndexError:
        print(i)
        second_attempt_failed.append(i)
        continue

third_attempt_failed = []

for i in second_attempt_failed:
    count = 0
    if count % 25 == 0:
        f = open("players_dictionary.pkl","wb")
        pickle.dump(d,f)
        f.close()
    try:
        d[i] = get_historical_prices(i)
        count += 1
    except IndexError:
        print(i)
        third_attempt_failed.append(i)
        continue

fourth_attempt_failed = []

for i in third_attempt_failed:
    count = 0
    if count % 25 == 0:
        f = open("players_dictionary.pkl","wb")
        pickle.dump(d,f)
        f.close()
    try:
        d[i] = get_historical_prices(i)
        count += 1
    except IndexError:
        print(i)
        fourth_attempt_failed.append(i)
        continue

f = open("players_dictionary.pkl","wb")
pickle.dump(d,f)
f.close()

players['Historical_Values'] = players['Profile_Link'].map(d)

players['Agent'] = players['Historical_Values'].apply(lambda row: row[len(row) - 1])

players['Historical_Values'] = players['Historical_Values'].apply(lambda row: row[:len(row) - 1])

players.to_pickle("Transfermarkt_Historicals.pkl")
