==> Final_Project/04.Get_Relevant_WhoScored_URLs.ipynb <==
import pandas as pd
import numpy as np

from copy import deepcopy
import unidecode

import pickle

players = pd.read_pickle("Players_Historicals_With_Injuries.pkl")

players.head(1)

whoscored = pd.read_csv("/Users/georgevarelas/Metis/Final Project/Code/scraper-whoscored-master MAIN/players_data/whoscored_data.txt", header = 0)


len(whoscored.columns)

newcols = ['Player']

for i in range(30):
    newcols.append(str(i))

whoscored.columns = newcols

whoscored = whoscored[['Player', "0"]]

def extract_name(url):
    if url.split("[\'")[1].split("-")[0] != "Undefined'":
        return url.split("[\'")[1].split("-")[0]
    else:
        return url.split('Show/')[1].split(':')[0][:-1].replace("-", " ")

whoscored['Name'] = whoscored['Player'].apply(extract_name)

whoscored.drop_duplicates(inplace = True)

whoscored_test = deepcopy(whoscored)

whoscored_test = whoscored_test[whoscored_test['Name'].isin(players['Name'].unique())]

whoscored_test[whoscored_test['Name'] == 'Kylian Mbappe']

not_in_whoscored = np.setdiff1d(players['Name'],whoscored_test['Name'])

not_in_whoscored

to_get_manually = ['Arturo Vidal', 'Christian Pulisic',  'David Alaba',\
                  'Frenkie de Jong', 'Heung min Son', 'Jadon Sancho', 'James Rodriguez', 'Javi Martinez','Jerome Boateng',\
                  'Joshua Kimmich', 'Kingsley Coman','Marco Reus', \
                  'Matthijs de Ligt', "N'Golo Kante", 'Philippe Coutinho', 'Robert Lewandowski', 'Thomas Muller', \
                  ]

links = ['/Players/25008/Show/Arturo-Vidal', '/Players/302692/Show/Christian-Pulisic', \
'/Players/39722/Show/David-Alaba', '/Players/279423/Show/Frenkie-de-Jong', '/Players/91909/Show/Son-Heung-Min', \
'/Players/346300/Show/Jadon-Sancho', '/Players/71182/Show/James-Rodríguez', '/Players/14091/Show/Javi-Martínez', \
'/Players/23589/Show/Jérôme-Boateng','/Players/283323/Show/Joshua-Kimmich', '/Players/125883/Show/Kingsley-Coman', \
'/Players/41330/Show/Marco-Reus', '/Players/320374/Show/Matthijs-de-Ligt', '/Players/114075/Show/N-Golo-Kanté', \
'/Players/80767/Show/Coutinho', '/Players/29400/Show/Robert-Lewandowski', \
'/Players/37099/Show/Thomas-Müller']

newlinks = []
for i in links:
    newlinks.append(
    ("{'" + i + ": " + '"[' + "'" + i.split('Show/')[1].replace("-", " ") + "-current'") 
    )

whoscored_test.drop('0', axis = 1, inplace = True)

to_append = []
for i in range(len(newlinks)):
    to_append.append([newlinks[i],to_get_manually[i]])

whoscored_test = whoscored_test.append(pd.DataFrame(to_append, columns=['Player','Name']),ignore_index=True)

whoscored_test['Player'] = whoscored_test['Player'].apply(lambda row: row.split(":")[0][2:])

whoscored_test['Player'] = whoscored_test['Player'].apply(lambda row: row[:-1] if row[-1] == "'" else row)

with open('whoscored_urls_narrow.txt', 'w') as f:
    for item in list(whoscored_test['Player']):
        f.write("%s\n" % item)

whoscored_test.to_pickle('whoscored_urls_transfermarkt_names.pkl')
