==> Final_Project/06.Whoscored_Cleaner.ipynb <==
import pandas as pd
import numpy as np

dirty = pd.read_csv('/Users/georgevarelas/Downloads/scraper-whoscored-master MAIN/players_historical_data/players_historical_data.csv')

for index, row in dirty.iterrows():
    if index%2 == 0:
        names.append(row)

names = dirty.iloc[::2]

values = dirty.iloc[1::2]

values = values.rename(columns={'Player Name': 'Season'})

#Seasons were stored as ratios in the csv file, need to convert back to interpretable numbers
d = {}
for i in range(2009, 2019):
    d[str(-i/(i+1))[:-7]] = str(i)

for i in range(2009, 2019):
    d[str(round(-i/(i+1), 9))] = str(i) 

newvals = []
for i in values['Season']:
    if abs(float(i)) < 1:
        newvals.append(d[i])
    else:
        newvals.append(i)

values.reset_index(inplace = True)

Season_1 = np.asarray(newvals)
values["Season"] = Season_1
values.drop('index', axis = 1, inplace = True)
values['Season'] = values['Season'].apply(lambda row: abs(int(row)))

names.reset_index(inplace = True)
values['Player Name'] = names['Player Name']

cols = values.columns.tolist()
cols = cols[-1:] + cols[:-1]
values = values[cols]

numerical_cols = ['Goals/90min',
       'Assists/90min', 'Yel/90min', 'Red/90min', 'SpG', 'PS%', 'Rating',
       'Tackles', 'Inter', 'Fouls (def)', 'Offsides', 'Clear', 'DrB (def)',
       'Blocks', 'DrB (off)', 'Fouled (off)', 'Off (off)', 'Disp (off)',
       'KeyP', 'AvgP', 'Crosses', 'LongB', 'ThrB', 'OutOfBox', 'SixYardBox',
       'PenaltyArea']

for i in numerical_cols:
    values[i] = values[i].replace('Undefined', 0)

values = values[['Player Name', 'Season', 'Team', 'Tournament', 'Goals/90min',
       'Assists/90min', 'Yel/90min', 'Red/90min', 'SpG', 'PS%', 'Rating',
       'Tackles', 'Inter', 'Fouls (def)', 'Offsides', 'Clear', 'DrB (def)',
       'Blocks', 'DrB (off)', 'Fouled (off)', 'Off (off)', 'Disp (off)',
       'KeyP', 'AvgP', 'Crosses', 'LongB', 'ThrB', 'OutOfBox', 'SixYardBox',
       'PenaltyArea']]

urls = pd.read_csv('/Users/georgevarelas/Downloads/scraper-whoscored-master MAIN/players_historical_data/historical_data_urls.rtf', header = 0)

urls = urls[6:]
urls = urls.iloc[1::2]

urls.reset_index(inplace = True)
urls.drop('index', axis = 1, inplace = True)

urls['URL'] = urls['URL'].apply(lambda row: 'Show'.join(row.split('History')))
values['URL'] = urls['URL']

values.to_pickle('Whoscored_Historicals.pkl')
