==> Final_Project/08.Get_Relevant_Fifa_Urls.ipynb <==
Extract FIFA urls of relevant players so we can scrape other attributes for the analysis

import pandas as pd
import unidecode

fifa_17_names = pd.read_csv("/Users/georgevarelas/Metis/Final Project/Fifa_Datasets/Fifa_17_names.csv")
fifa_17 = pd.read_csv('/Users/georgevarelas/Metis/Final Project/Fifa_Datasets/Fifa_17.csv')
df = pd.read_pickle('/Users/georgevarelas/Downloads/scraper-whoscored-master MAIN/TRANSFERMARKT_WHOSCORED_MERGE.pkl')

fifa_17 = pd.merge(fifa_17, fifa_17_names, on = ['Name'])
fifa_17 = fifa_17[['Name', 'Age','Club', 'url']]
fifa_17['Name'] = fifa_17['Name'].apply(lambda row: unidecode.unidecode(row).replace("-", " "))
fifa_17['Age'] = fifa_17['Age'].apply(lambda row: float(row))

df['Age'] = df['Age'].apply(lambda row: float(row))
df = df.groupby(['Name', 'Season'])[['Age']].mean().reset_index()

not_in_17 = []
for i in df['Name'].unique():
    if i not in fifa_17['Name'].unique():
        if i.split(' ')[-1] in fifa_17['LastName'].unique():
            lastname_in_17.append(i)
        else:
            not_in_17.append(i)


def extract_last_name(row):
    if len(row.split(' ')) == 1:
        return (row)
    else:
        return row.split(' ')[-1]

not_in_17 = []
for i in df['Name'].unique():
    if i not in fifa_17['Name'].unique():
            not_in_17.append(i)

len(df['Name'].unique())

len(not_in_17)

fifa_17 = fifa_17[['Name', 'url']][fifa_17['Name'].isin(df['Name'].unique())].drop_duplicates()

name = ['Alvaro Morata', 'Andres Iniesta', 'Callum Hudson Odoi', 'Cesar Azpilicueta', 'Daniel Carvajal', 'Gerard Pique', \
       'Heung min Son', 'Jadon Sancho', 'Mousa Dembele', 'Philippe Coutinho', 'Saul Niguez', 'Victor Lindelof', \
       'Vinicius Junior', 'Alexandre Pato', 'Arthur', 'Chicharito', 'Hulk', 'Kylian Mbappe', 'Lucas Moura', \
       'Scott McTominay', 'Timothy Weah' ]

url_old = ['https://www.fifaindex.com/player/201153/morata/', 'https://www.fifaindex.com/player/41/iniesta/', 'https://www.fifaindex.com/player/240740/callum-hudson-odoi/', \
      'https://www.fifaindex.com/player/184432/azpilicueta/', 'https://www.fifaindex.com/player/204963/carvajal/', 'https://www.fifaindex.com/player/152729/piqu%C3%A9/', \
      'https://www.fifaindex.com/player/200104/heung-min-son/', 'https://www.fifaindex.com/player/233049/jadon-sancho/', 'https://www.fifaindex.com/player/162240/moussa-demb%C3%A9l%C3%A9/', \
      'https://www.fifaindex.com/player/189242/coutinho/', 'https://www.fifaindex.com/player/208421/sa%C3%BAl/', \
      'https://www.fifaindex.com/player/221660/victor-lindel%C3%B6f/', 'https://www.fifaindex.com/player/238794/vin%C3%ADcius-j%C3%BAnior/', \
      'https://www.fifaindex.com/player/180175/alexandre-pato/fifa19_326/', 'https://www.fifaindex.com/player/230658/arthur/', \
      'https://www.fifaindex.com/player/178224/javier-hern%C3%A1ndez/', 'https://www.fifaindex.com/player/189362/hulk/',  \
      'https://www.fifaindex.com/player/231747/kylian-mbapp%C3%A9/', 'https://www.fifaindex.com/player/200949/lucas-moura/', \
       'https://www.fifaindex.com/player/237238/scott-mctominay/', 'https://www.fifaindex.com/player/241496/timothy-weah/' 
      ]
url = [i.split('.com')[1] for i in url_old]

add_to_17 = pd.DataFrame()
add_to_17['Name'] = name
add_to_17['url'] = url

fifa_17 = fifa_17.reset_index().drop('index', axis = 1)
fifa_17 = fifa_17.append(add_to_17, ignore_index=True)

fifa_17.to_csv('Relevant_Fifa_Urls')

lastname_in = ['Aaron Wan Bissaka',
 'Agustin Almendra',
 'Alexandre Pato',
 'Andrew Gravillon',
 'Arthur',
 'Bruno Guimaraes',
 'Chicharito',
 'Domagoj Vida',
 'Eder Militao',
 'Eljif Elmas',
 'Elvis Rexhbecaj',
 'Enis Bardhi',
 'Enzo Loiodice',
 'Eran Zahavi',
 'Erling Braut Haland',
 'Ethan Ampadu',
 'Everton',
 'Fabian Balbuena',
 'Fabio Depaoli',
 'Fagner',
 'Felix Eboa Eboa',
 'Gabriel Paulista',
 'Gervinho',
 'Giorgian de Arrascaeta',
 'Gonzalo Maroni',
 'Hakan Calhanoglu',
 'Hulk',
 'Jordan Siebatcheu',
 'Juan Cazares',
 'Jules Kounde',
 'Junior Firpo',
 'Kelvin Amian',
 'Keno',
 'Kylian Mbappe',
 'Leo Jaba',
 'Leonardo Koutris',
 'Lincoln',
 'Linton Maina',
 'Loren Moron',
 'Luan',
 'Lucas Moura',
 'Lucas Paqueta',
 'Lucas Verissimo',
 'Marco Richter',
 'Marin Pongracic',
 'Mario Hermoso',
 'Martin Aguirregabiria',
 'Maycon',
 'Nahitan Nandez',
 'Nemanja Gudelj',
 'Nikola Milenkovic',
 'Nikola Vlasic',
 'Noussair Mazraoui',
 'Odion Ighalo',
 'Pietro Pellegri',
 'Ramon Abila',
 'Ricardo Goulart',
 'Richarlison',
 'Rodrigo Becao',
 'Rodrygo',
 'Roger Marti',
 'Rogerio',
 'Scott McTominay',
 'Timothy Weah',
 'Tomas Chancalay',
 'Trezeguet',
 'Vitor Hugo',
 'Walter Kannemann',
 'Wendel',
 'Weston McKennie',
 'Willem Geubbels',
 'Yacine Adli']

not_in = ['Alejandro Grimaldo',
 'Aleksandr Erokhin',
 'Aleksandr Golovin',
 'Aleksandr Kokorin',
 'Aleksey Miranchuk',
 'Alex Granell',
 'Alex Teixeira',
 'Alvaro Gonzalez',
 'Alvaro Morata',
 'Alvaro Odriozola',
 'Amadou Haidara',
 'Andres Iniesta',
 'Asier Illarramendi',
 'Ayrton Lucas',
 'Boubakary Soumare',
 'Brais Mendez',
 'Callum Hudson Odoi',
 'Cedric Soares',
 'Cesar Azpilicueta',
 'Cristhian Stuani',
 'Daniel Carvajal',
 'Didac Vila',
 'Dyego Sousa',
 'Everton Ribeiro',
 'Fabian Ruiz',
 'Federico Valverde',
 'Franck Kessie',
 'Gabriel Boschilia',
 'Gerard Deulofeu',
 'Gerard Pique',
 'Heung min Son',
 'Ibrahima Konate',
 'Igor Zubeldia',
 'Iker Muniain',
 'Inaki Williams',
 'Ivan Marcano',
 'Jadon Sancho',
 'Javier Manquillo',
 'Jesus Vallejo',
 'Jose Campana',
 'Jose Gaya',
 'Jose Gimenez',
 'Jose Luis Morales',
 'Jose Luis Palomino',
 'Joseba Zaldua',
 'Luiz Araujo',
 'Marc Bartra',
 'Markel Susaeta',
 'Martin Montoya',
 'Mbaye Diagne',
 'Mikel Oyarzabal',
 'Mousa Dembele',
 'Moussa Diaby',
 'Munir El Haddadi',
 'Musa Barrow',
 'Nemanja Maksimovic',
 'Oghenekaro Etebo',
 'Ole Selnaes',
 'Orel Mangala',
 'Phil Foden',
 'Philippe Coutinho',
 'Rafael Leao',
 'Raul Silva',
 'Reiss Nelson',
 'Rodrigo Caio',
 'Rodrigo Dourado',
 'Rodrigo de Paul',
 'Roger Guedes',
 'Rony Lopes',
 'Ruben Dias',
 'Sandro Ramirez',
 'Saul Niguez',
 'Sergio Canales',
 'Thiago Mendes',
 'Tiquinho Soares',
 'Toni Lato',
 'Unai Nunez',
 'Vicente Iborra',
 'Victor Camarasa',
 'Victor Laguardia',
 'Victor Lindelof',
 'Vinicius Junior',
 'Wilfred Ndidi',
 'Yeray Alvarez',
 'Yerry Mina',
 'Yuri Gazinskiy']
