==> Final_Project/09.Fifa_Scrape.ipynb <==
from bs4 import BeautifulSoup
#import urllib2
import requests
import pandas as pd
import pickle

from __future__ import print_function, division
import sys
from bs4 import BeautifulSoup
import requests

requests.__path__

import pandas as pd
import pickle

df = pd.read_csv('/Users/georgevarelas/Relevant_Fifa_Urls')

def fifa_scrape(link, fifa_version):
    
    if fifa_version == 19:
        url = "https://www.fifaindex.com" + link
    elif fifa_version == 18:
        url = "https://www.fifaindex.com" + link + "fifa18_278/"
    elif fifa_version == 17:
        url = "https://www.fifaindex.com" + link + "fifa17_173/"
    elif fifa_version == 16:
        url = "https://www.fifaindex.com" + link + "fifa16_73/"    
    elif fifa_version == 15:
        url = "https://www.fifaindex.com" + link + "fifa15_14/"
    elif fifa_version == 14:
        url = "https://www.fifaindex.com" + link + "fifa14_13/"
    elif fifa_version == 13:
        url = "https://www.fifaindex.com" + link + "fifa13_10/"
        
    user_agent = {'User-agent': 'Mozilla/5.0'}
    page  = requests.get(url, headers = user_agent)
    html = page.content
    soup = BeautifulSoup(html, "lxml")
    
    Link = []
    Height = []
    Weight = []
    Wage = []
    Contract = []
    Age = []
    
    Link.append(link)
    
    right_table=soup.findAll('div', class_='card-body')
    
    attr = right_table[1].findAll('span')
    
    Height.append(attr[1].text)
    Weight.append(attr[5].text)
    Age.append(attr[8].text)
    Wage.append(attr[-3].text)
    

    table_strings = [str(i) for i in right_table]
    
    contract_index = [i for i, s in enumerate(table_strings) if 'Contract' in s][0]
    Contract.append(right_table[contract_index].findAll('span')[-1].text)
        
    physical_index = [i for i, s in enumerate(table_strings[1:]) if 'Acceleration' in s][0]+1
    physical = right_table[physical_index].findAll('span')
    physical_data = [i.text for i in physical][::2]
    
    intermediate = Link + Height + Weight +  Age + Wage + Contract + physical_data
        
    result = intermediate[:6] + [s for s in intermediate[6:] if int(s) >= 15]
    
    return result
    

df = df[df.url != '/player/180175/alexandre-pato/fifa19_326/']

df.reset_index(inplace = True)

fifa_19 = []
fifa_18 = []
fifa_17 = []
fifa_16 = []
fifa_15 = []
fifa_14 = []
fifa_13 = []

for i in range(len(df)):
    fifa_19.append(fifa_scrape(df['url'][i], 19))
    
fifa_19_df = pd.DataFrame(fifa_19, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

for i in range(len(df)):
    fifa_18.append(fifa_scrape(df['url'][i], 18))
    
fifa_18_df = pd.DataFrame(fifa_18, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

for i in range(len(df)):
    fifa_17.append(fifa_scrape(df['url'][i], 17))
    
fifa_17_df = pd.DataFrame(fifa_17, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

for i in range(len(df)):
    fifa_16.append(fifa_scrape(df['url'][i], 16))
    
shouldnt_be_in_16_indices = [i for i, s in enumerate(fifa_16) if len(s) > 18]

new_fifa_16 = []
for i in range(len(fifa_16)):
    if i not in shouldnt_be_in_16_indices:
        new_fifa_16.append(fifa_16[i])

fifa_16_df = pd.DataFrame(new_fifa_16, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

for i in range(len(df)):
    fifa_15.append(fifa_scrape(df['url'][i], 15))
    
shouldnt_be_in_15_indices = [i for i, s in enumerate(fifa_15) if len(s) > 18]

new_fifa_15 = []
for i in range(len(fifa_15)):
    if i not in shouldnt_be_in_15_indices:
        new_fifa_15.append(fifa_15[i])
        
fifa_15_df = pd.DataFrame(new_fifa_15, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

for i in range(len(df)):
    fifa_14.append(fifa_scrape(df['url'][i], 14))

shouldnt_be_in_14_indices = [i for i, s in enumerate(fifa_14) if len(s) > 18]

new_fifa_14 = []
for i in range(len(fifa_14)):
    if i not in shouldnt_be_in_14_indices:
        new_fifa_14.append(fifa_14[i])
        
fifa_14_df = pd.DataFrame(new_fifa_14, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

for i in range(len(df)):
    fifa_13.append(fifa_scrape(df['url'][i], 13))
    
shouldnt_be_in_13_indices = [i for i, s in enumerate(fifa_13) if len(s) > 18]

new_fifa_13 = []
for i in range(len(fifa_13)):
    if i not in shouldnt_be_in_13_indices:
        new_fifa_13.append(fifa_13[i])
        
fifa_13_df = pd.DataFrame(new_fifa_13, columns=['url','Height','Weight','Age','Wage','Contract_Expiry','Acceleration', 'Stamina', 'Strength', 'Balance', 'Sprint Speed', \
                          'Agility', 'Jumping'])

fifa_19_df.to_pickle('fifa_19_df.pkl')
fifa_18_df.to_pickle('fifa_18_df.pkl')
fifa_17_df.to_pickle('fifa_17_df.pkl')
fifa_16_df.to_pickle('fifa_16_df.pkl')
fifa_15_df.to_pickle('fifa_15_df.pkl')
fifa_14_df.to_pickle('fifa_14_df.pkl')
fifa_13_df.to_pickle('fifa_13_df.pkl')
