==> Final_Project/10.Fifa_Join_.ipynb <==
import pandas as pd
import pickle

df = pd.read_csv('/Users/georgevarelas/Relevant_Fifa_Urls')
main = pd.read_pickle('/Users/georgevarelas/Downloads/scraper-whoscored-master MAIN/TRANSFERMARKT_WHOSCORED_MERGE.pkl')

fifa_13_df = pd.read_pickle('/Users/georgevarelas/fifa_13_df.pkl')
fifa_14_df = pd.read_pickle('/Users/georgevarelas/fifa_14_df.pkl')
fifa_15_df = pd.read_pickle('/Users/georgevarelas/fifa_15_df.pkl')
fifa_16_df = pd.read_pickle('/Users/georgevarelas/fifa_16_df.pkl')
fifa_17_df = pd.read_pickle('/Users/georgevarelas/fifa_17_df.pkl')
fifa_18_df = pd.read_pickle('/Users/georgevarelas/fifa_18_df.pkl')
fifa_19_df = pd.read_pickle('/Users/georgevarelas/fifa_19_df.pkl')

main = main.dropna()

main['Market_Value'] = main['Market_Value'].apply(lambda row: float(row))
main['Age'] = main['Age'].apply(lambda row: float(row))

fifa_13_df['Season'] = 2012
fifa_14_df['Season'] = 2013
fifa_15_df['Season'] = 2014
fifa_16_df['Season'] = 2015
fifa_17_df['Season'] = 2016
fifa_18_df['Season'] = 2017
fifa_19_df['Season'] = 2018

to_drop_new = ['Aggression', 'Reactions', 'Att_Position', 'Interceptions', 'Vision',\
               'Composure', 'Acceleration', 'Stamina', 'Strength', 'Balance',\
               'Sprint Speed', 'Agility', 'Jumping']
to_drop_old = ['Aggression', 'Reactions', 'Att_Position', 'Interceptions', 'Vision',\
                'Acceleration', 'Stamina', 'Strength', 'Balance',\
               'Sprint Speed', 'Agility', 'Jumping']

fifa_19_df = fifa_19_df.drop(columns = to_drop_new)
fifa_18_df = fifa_18_df.drop(columns = to_drop_new)
fifa_17_df = fifa_17_df.drop(columns = to_drop_new)

fifa_16_df = fifa_16_df.drop(columns = to_drop_old)
fifa_15_df = fifa_15_df.drop(columns = to_drop_old)
fifa_14_df = fifa_14_df.drop(columns = to_drop_old)
fifa_13_df = fifa_13_df.drop(columns = to_drop_old)


df = df.iloc[:, 1:]

df['times'] = 7

test = df.loc[df.index.repeat(df.times)].reset_index(drop = True)
test['Season']  = [(i % 7) + 2012 for i in range(len(test))]
test = test.drop(columns = 'times')
test = pd.merge(test, fifa_19_df, how = 'outer', on = ['url', 'Season'])

fifa = pd.concat([fifa_19_df, fifa_18_df, fifa_17_df, fifa_16_df, fifa_15_df, fifa_14_df, fifa_13_df])

fifa_final = pd.merge(test, fifa, how = 'outer', on = ['url', 'Season'])

#Moussa Dembele was a duplicate name, ended up scraping the wrong one
fifa_final = fifa_final[fifa_final.Name != 'Moussa Dembele']
fifa_final = fifa_final.dropna()
fifa_final = fifa_final.reset_index(drop = True)
fifa_final = fifa_final.drop(fifa_final.index[[6236, 6237, 6238, 6239, 6240]])

fifa_final['Height'] = fifa_final['Height'].apply(lambda row: float(row.split(' ')[0]))
fifa_final['Weight'] = fifa_final['Weight'].apply(lambda row: float(row.split(' ')[0]))

fifa_final = fifa_final[fifa_final.Name != 'Hulk']

fifa_final['Wage'] = fifa_final['Wage'].apply(lambda row: (row[1:].split('.')[0] + '000'))
fifa_final['Wage'] = fifa_final['Wage'].apply(lambda row: float(row))

fifa_final.to_pickle('fifa_final.pkl')
