==> Final_Project/11.Final_Modelling.ipynb <==
import numpy as np
import pandas as pd
import pickle
import warnings

import statsmodels.api as sm
import statsmodels.formula.api as smf
import patsy
 
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression, Ridge, RidgeCV, Lasso, LassoCV  #ordinary linear regression + w/ ridge regularization
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score
from sklearn import feature_selection as f_select
from sklearn.metrics import r2_score
from sklearn.linear_model import ElasticNetCV

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import RandomizedSearchCV

warnings.filterwarnings('ignore')

df = pd.read_pickle('/Users/georgevarelas/Downloads/scraper-whoscored-master MAIN/TRANSFERMARKT_WHOSCORED_MERGE.pkl')
fifa = pd.read_pickle('/Users/georgevarelas/fifa_final.pkl')

df.shape

df = df.dropna()

#reorder columns
df = df[['Name', 'Market_Value',  'Position','Season', 'Age', 'Team', 'Tournament','Games_Missed', 'Goals/90min', 'Assists/90min',
       'Yel/90min', 'Red/90min', 'SpG', 'PS%', 'Rating', 'Tackles', 'Inter',
       'Fouls (def)', 'Offsides', 'Clear', 'DrB (def)', 'Blocks', 'DrB (off)',
       'Fouled (off)', 'Off (off)', 'Disp (off)', 'KeyP', 'AvgP', 'Crosses',
       'LongB', 'ThrB', 'OutOfBox', 'SixYardBox', 'PenaltyArea', 'URL']]

#change Market_Value and Age to float:
"""df['Market_Value'] = df['Market_Value'].apply(lambda row: float(row))
df['Age'] = df['Age'].apply(lambda row: float(row))"""

df.shape

df = pd.merge(df, fifa, on = ['Name', 'Season'])

#get some weird age differences from fifa dataset and transfermarkt. Get rid of any rows with large age differences

df['Age_y'] = df['Age_y'].apply(lambda row: float(row))
df['Age_Diff'] = df['Age_x'] - df['Age_y']
#drop rows where player ages are very mismatched
df = df.drop(df[abs(df['Age_Diff']) >= 3.0 ].index)
df = df.drop(columns = ['Age_Diff', 'Age_x'])
df = df.rename(index=str, columns={"Age_y": "Age"})

#Make Years to Contract_Expiry
df['Contract_Expiry'] = df['Contract_Expiry'].apply(lambda row: float(row))
df['Seasons_Left_On_Contract'] = df['Contract_Expiry'] - df['Season']

df['Team'] = df['Team'].apply(lambda x: x.split(' U21')[0])

#pickle this so we can try to get google search trends
df.to_pickle('final_players_dataset.pkl')

#create columns specific to club/intl tournaments so we can groupby later and preserve information

international_club = ['UEL', 'UCL']
club_league = ['BSA', 'RPL','CSl','ISA','GB','FL1','EPL','NE','SLL','APD','PLN', 'TS', 'SA', 'UMLS','EC', 'GB2']
international = ['UNL','WC', 'UEC', 'ICA','ICC', 'ACoN','IEU']

#Ranking of most powerful clubs/international teams per season
forbes_2018 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Bayern Munich',\
              'Manchester City', 'Chelsea', 'Arsenal', 'Liverpool', 'Tottenham', \
              'Juventus', 'Paris Saint-Germain', 'Atletico Madrid', 'Borussia Dortmund', 'Schalke 04', \
              'Inter', 'Roma', 'West Ham', 'AC Milan', 'Everton', 'Newcastle United']

forbes_2017 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Bayern Munich','Manchester City', \
              'Chelsea', 'Arsenal', 'Liverpool', 'Juventus', 'Tottenham', \
               'Paris Saint-Germain','Borussia Dortmund', 'Atletico Madrid','West Ham','Schalke 04', \
              'Inter', 'Roma', 'AC Milan', 'Leicester', 'Napoli']

forbes_2016 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Bayern Munich','Manchester City',\
               'Chelsea', 'Arsenal', 'Liverpool', 'Juventus','Tottenham', \
               'Paris Saint-Germain', 'Borussia Dortmund',  'AC Milan', 'Atletico Madrid', 'West Ham', \
              'Schalke 04','Inter', 'Roma',  'Leicester', 'Napoli']

forbes_2015 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Bayern Munich', 'Arsenal',\
              'Manchester City', 'Chelsea',  'Liverpool', 'Juventus', 'Tottenham', \
               'Borussia Dortmund','AC Milan','Paris Saint-Germain',  'Schalke 04', 'Atletico Madrid',  \
              'Inter', 'Roma', 'West Ham',  'Newcastle United', 'Napoli']

forbes_2014 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Bayern Munich', 'Manchester City',\
               'Chelsea', 'Arsenal', 'Liverpool','Juventus', 'AC Milan', \
               'Borussia Dortmund','Paris Saint-Germain', 'Tottenham', 'Schalke 04', \
              'Inter', 'Atletico Madrid', 'Galatasaray', 'West Ham',  'Newcastle United', 'Napoli']

forbes_2013 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Bayern Munich','Arsenal',\
              'Chelsea', 'Manchester City', 'AC Milan', 'Juventus', 'Liverpool',\
               'Borussia Dortmund', 'Schalke 04','Tottenham', 'Inter', 'Paris Saint-Germain',\
               'Galatasaray', 'Atletico Madrid', 'Roma',  'Hamburger SV', 'Napoli']

forbes_2012 = ['Real Madrid', 'Manchester United', 'Barcelona', 'Arsenal', 'Bayern Munich'\
              'AC Milan', 'Chelsea', 'Juventus', 'Manchester City', 'Liverpool', \
               'Tottenham','Schalke 04', 'Borussia Dortmund', 'Inter', 'Lyon',  \
               'Corinthians', 'Napoli', 'Hamburger SV', 'Marseille', 'Newcastle United']

world_2018 = ['Belgium', 'France', 'Brazil', 'England', 'Croatia', \
              'Uruguay', 'Portugal', 'Switzerland', 'Spain', 'Denmark', \
              'Argentina', 'Colombia', 'Germany', 'Sweden', 'Chile', \
              'Netherlands', 'Italy', 'Mexico', 'Wales', 'Poland']

world_2017 = ['Belgium', 'France', 'Brazil','Croatia','England',\
             'Uruguay', 'Portugal', 'Switzerland', 'Spain', 'Denmark', \
             'Argentina', 'Colombia','Sweden', 'Chile','Netherlands',\
              'Germany', 'Italy', 'Mexico', 'Wales', 'Poland']

world_2016 =  ['Germany', 'Brazil', 'Portugal', 'Argentina','Belgium',\
              'Spain', 'Poland', 'Switzerland', 'France', 'Chile', \
              'Peru', 'Denmark', 'Colombia', 'Italy', 'England', \
              'Mexico', 'Croatia', 'Sweden', 'Wales', 'Netherlands']

world_2015 = ['Germany', 'Brazil', 'Argentina','Belgium', 'Chile',\
             'Colombia', 'Portugal','France','Uruguay', 'Spain', \
             'Switzerland', 'Wales', 'England', 'Croatia', 'Poland',\
             'Italy', 'Costa Rica', 'Mexico', 'Peru', 'Ecuador']

world_2014 = ['Germany', 'Spain', 'Argentina','Belgium', 'Chile', \
              'Brazil','Colombia', 'Portugal', 'England', 'Austria', \
              'Uruguay', 'Switzerland', 'Ecuador', 'Netherlands', 'Italy', \
              'Romania', 'Wales', 'Croatia', 'Ivory Coast', 'Hungary']

world_2013 = ['Germany', 'Colombia', 'Argentina','Belgium', 'Netherlands',\
              'Brazil','Portugal','France','Spain','Uruguay',\
              'England', 'Italy', 'Switzerland', 'Chile', 'Romania'\
              'Costa Rica', 'Czech Republic', 'Algeria', 'Croatia', 'Mexico']


world_2012 = ['Germany', 'Colombia', 'Argentina','Spain', 'Portugal',\
              'Uruguay', 'Italy', 'Switzerland', 'Netherlands', 'Brazil', \
              'Belgium', 'Greece', 'England', 'USA', 'Chile', \
              'Croatia', 'Ivory Coast', 'Ukraine', 'Bosnia and Herzegovina', 'France']               

def get_club_ranking(row, forbes_list):
    if row in forbes_list[:5]:
        return 5
    elif row in forbes_list[5:10]:
        return 4
    elif row in forbes_list[10:15]:
        return 3
    elif row in forbes_list[15:20]:
        return 2
    else:
        return 1

df['Forbes_Rank_18'] = df[df['Season'] == 2018]['Team'].apply(lambda x: get_club_ranking(x, forbes_2018) if x in forbes_2018 else get_club_ranking(x, world_2018))
df['Forbes_Rank_17'] = df[df['Season'] == 2017]['Team'].apply(lambda x: get_club_ranking(x, forbes_2017) if x in forbes_2017 else get_club_ranking(x, world_2017))
df['Forbes_Rank_16'] = df[df['Season'] == 2016]['Team'].apply(lambda x: get_club_ranking(x, forbes_2016) if x in forbes_2016 else get_club_ranking(x, world_2016))
df['Forbes_Rank_15'] = df[df['Season'] == 2015]['Team'].apply(lambda x: get_club_ranking(x, forbes_2015) if x in forbes_2015 else get_club_ranking(x, world_2015))
df['Forbes_Rank_14'] = df[df['Season'] == 2014]['Team'].apply(lambda x: get_club_ranking(x, forbes_2014) if x in forbes_2014 else get_club_ranking(x, world_2014))
df['Forbes_Rank_13'] = df[df['Season'] == 2013]['Team'].apply(lambda x: get_club_ranking(x, forbes_2013) if x in forbes_2013 else get_club_ranking(x, world_2013))
df['Forbes_Rank_12'] = df[df['Season'] == 2012]['Team'].apply(lambda x: get_club_ranking(x, forbes_2012) if x in forbes_2012 else get_club_ranking(x, world_2012))

df.iloc[:, -7:] = df.iloc[:, -7:].fillna(0)
df['Forbes_Rank'] = df['Forbes_Rank_18'] + df['Forbes_Rank_17'] + df['Forbes_Rank_16'] + df['Forbes_Rank_15'] \
+ df['Forbes_Rank_14'] + df['Forbes_Rank_13'] + df['Forbes_Rank_12']


df = df.iloc[:, np.r_[0:len(df.columns)-8, -1]]

df.columns

grouped = df.groupby(['Name', 'Season', 'Position'])[['Market_Value', 
       'Games_Missed', 'Goals/90min', 'Assists/90min', 'Yel/90min',
       'Red/90min', 'SpG', 'PS%', 'Rating', 'Tackles', 'Inter', 'Fouls (def)',
       'Offsides', 'Clear', 'DrB (def)', 'Blocks', 'DrB (off)', 'Fouled (off)',
       'Off (off)', 'Disp (off)', 'KeyP', 'AvgP', 'Crosses', 'LongB', 'ThrB',
       'OutOfBox', 'SixYardBox', 'PenaltyArea', 'Height',
       'Weight', 'Age', 'Wage', 'Seasons_Left_On_Contract', 'Forbes_Rank']].mean().reset_index()

#add lag_value as a column
grouped['Lag_Value'] = grouped.groupby('Name')['Market_Value'].shift(1)

#reorder columns so we can see lagged value
cols = grouped.columns.tolist()
cols
cols = cols[:4] + cols[-1:] + cols[4:-1]
grouped = grouped[cols]

#drop whoscored rating
grouped = grouped.drop('Rating', axis = 1)

#Lag the game stats & other metrics also
lag_columns = list(grouped.columns[5:])
for i in lag_columns:
    grouped[i] =  grouped.groupby('Name')[i].shift(1)

#rename lagged columns to _Lag
newcols = list(grouped.columns[:5]) + [i + "_Lag" for i in lag_columns]
grouped.columns = newcols

# Feature Engineering

#### Dummies

Forward = ['Centre-Forward', 'Second Striker']
Winger = ['Left Winger', 'Right Winger']
Central_Midfielder = ['Central Midfield', 'Right Midfield', 'Left Midfield']
Defensive_Midfielder = ['Defensive Midfield']
Attacking_Midfielder = ['Attacking Midfield']
Defender = ['Centre-Back']
Wing_Back = ['Right-Back', 'Left-Back']

grouped["Forward"] = grouped['Position'].apply(lambda row: 1 if row in (Forward) else 0)
grouped["Winger"] = grouped['Position'].apply(lambda row: 1 if row in (Winger) else 0)
grouped["Central_Midfielder"] = grouped['Position'].apply(lambda row: 1 if row in (Central_Midfielder) else 0)
grouped["Defensive_Midfielder"] = grouped['Position'].apply(lambda row: 1 if row in (Defensive_Midfielder) else 0)
grouped["Attacking_Midfielder"] = grouped['Position'].apply(lambda row: 1 if row in (Attacking_Midfielder) else 0)
grouped["Defender"] = grouped['Position'].apply(lambda row: 1 if row in (Defender) else 0)
grouped['Wing_Back'] = grouped['Position'].apply(lambda row: 1 if row in (Wing_Back) else 0)

#### Interactions

forward_interactions = ['Goals/90min_Lag', 'Assists/90min_Lag', 'DrB (off)_Lag', 'Off (off)_Lag', 'Disp (off)_Lag']
winger_interactions = ['Goals/90min_Lag', 'Assists/90min_Lag', 'DrB (off)_Lag', 'Off (off)_Lag', 'Disp (off)_Lag', 'Crosses_Lag']
central_midfielder_interactions = ['PS%_Lag', 'KeyP_Lag', 'Tackles_Lag', 'DrB (off)_Lag', 'Disp (off)_Lag', 'Inter_Lag']
defensive_midfielder_interactions = ['Tackles_Lag', 'Inter_Lag', 'Blocks_Lag', 'PS%_Lag', 'Fouls (def)_Lag', 'DrB (def)_Lag']
attacking_midfielder_interactions = ['Goals/90min_Lag', 'Assists/90min_Lag', 'DrB (off)_Lag', 'KeyP_Lag', 'PS%_Lag']
defender_interactions = ['Tackles_Lag', 'Inter_Lag', 'Blocks_Lag', 'Fouls (def)_Lag', 'DrB (def)_Lag']
wing_back_interactions = ['Tackles_Lag', 'Inter_Lag', 'Blocks_Lag', 'Fouls (def)_Lag', 'DrB (def)_Lag', 'Assists/90min_Lag', 'DrB (off)_Lag',\
                         'Disp (off)_Lag', 'Crosses_Lag']

def generate_interactions(df, position, position_interactions):
    for i in position_interactions:
        df[position+'_'+i] = df[i] * df[position]
    return df

positions = ['Forward', 'Winger', 'Central_Midfielder', 'Defensive_Midfielder', 'Attacking_Midfielder', 'Defender', 'Wing_Back']
interactions = [forward_interactions, winger_interactions, central_midfielder_interactions , defensive_midfielder_interactions,\
               attacking_midfielder_interactions, defender_interactions, wing_back_interactions]

#generate interactions and add them to dataframe
for i, j in list(zip(positions, interactions)):
    grouped = generate_interactions(grouped, i, j)

#### Lagged Features

lag_3_columns = list(grouped.columns[4:])
for i in lag_3_columns:
    grouped[i + "_3"] =  grouped.groupby('Name')[i].rolling(window = 2).mean().reset_index(drop = True)

len(grouped[grouped['Season'] <= 2017])/len(grouped)

# Modelling

#Drop NANs for Linear Regression

X_train_1 = grouped[grouped['Season'] <= 2017].dropna().iloc[:, np.r_[1, 4:len(grouped.columns)]]
X_test = grouped[grouped['Season'] == 2018].dropna().iloc[:, np.r_[1, 4:len(grouped.columns)]]
y_train_1 = grouped[grouped['Season'] <= 2017].dropna()['Market_Value']
y_test = grouped[grouped['Season'] == 2018].dropna()['Market_Value']

X_train, X_val, y_train, y_val = train_test_split(X_train_1, y_train_1, test_size=.15, random_state=5)

#FOR LASSO AND RIDGE
def mae(y_true, y_pred):
    return np.mean(np.abs(y_pred - y_true))

std = StandardScaler()
std.fit(X_train.values)

## Scale the Predictors on both the train and test set
X_tr = std.transform(X_train.values)
X_val_sc = std.transform(X_val.values)
X_test_sc = std.transform(X_test.values)

# Run the cross validation, find the best alpha, refit the model on all the data with that alpha
alphavec = 10**np.linspace(-2,2,200)


X_train_1_RF = grouped[grouped['Season'] <= 2017].fillna(-600).iloc[:, np.r_[1, 4:len(grouped.columns)]]
X_test_RF = grouped[grouped['Season'] == 2018].fillna(-600).iloc[:, np.r_[1, 4:len(grouped.columns)]]
y_train_1_RF = grouped[grouped['Season'] <= 2017].fillna(-600)['Market_Value']
y_test_RF = grouped[grouped['Season'] == 2018].fillna(-600)['Market_Value']

X_train_RF, X_val_RF, y_train_RF, y_val_RF = train_test_split(X_train_1_RF, y_train_1_RF, test_size=.15, random_state=5)


### Linear Regression

lm = LinearRegression()
#fit the training set
lm.fit(X_train.iloc[:, 3:33], y_train)
#get score for training and validation set
print('Training set R^2 = {}'.format(lm.score(X_train.iloc[:, 3:33], y_train)))
print('Validation set R^2 = {}'.format(lm.score(X_val.iloc[:, 3:33], y_val)))

#Baseline with just old market values
#LR with all columns
lm = LinearRegression()
#fit the training set
lm.fit(X_train.iloc[:,0:2], y_train)
#get score for training and validation set
print('Training set R^2 = {}'.format(lm.score(X_train.iloc[:,0:2], y_train)))
print('Validation set R^2 = {}'.format(lm.score(X_val.iloc[:,0:2], y_val)))

#LR with all columns
lm = LinearRegression()
#fit the training set
lm.fit(X_train, y_train)
#get score for training and validation set
print('Training set R^2 = {}'.format(lm.score(X_train, y_train)))
print('Validation set R^2 = {}'.format(lm.score(X_val, y_val)))

#### Lasso CV

lasso_model = LassoCV(alphas = alphavec, cv=5)
lasso_model.fit(X_tr, y_train)

print('Lasso model alpha = {}'.format(lasso_model.alpha_))

# Make predictions on the validation set using the new model
val_set_pred = lasso_model.predict(X_val_sc)

print('Lasso model MAE = {}'.format(mae(y_val, val_set_pred)))
print('Lasso Validation R^2 = {}'.format(r2_score(y_val, val_set_pred)))

#### Ridge CV

ridge_model = RidgeCV(alphas = alphavec, cv=5)
ridge_model.fit(X_tr, y_train)

print('Ridge model alpha = {}'.format(ridge_model.alpha_))

# Make predictions on the validation set using the new model
ridge_val_set_pred = ridge_model.predict(X_val_sc)

print('Ridge model MAE = {}'.format(mae(y_val, ridge_val_set_pred)))
print('Ridge Validation R^2 = {}'.format(r2_score(y_val, ridge_val_set_pred)))

resids = y_val - ridge_val_set_pred
residuals = pd.DataFrame({'Residuals':resids, 'Predicted Values': ridge_val_set_pred})

pd.set_option('display.float_format', '{:,.1f}'.format)

residuals.reset_index(inplace = True)
residuals['Name'] = residuals['index'].apply(lambda row: grouped.iloc[row, [0]])
residuals['Market_Value'] = residuals['index'].apply(lambda row: grouped.iloc[row, 3])
residuals['Season'] = residuals['index'].apply(lambda row: grouped.iloc[row, 1])
residuals['Position'] = residuals['index'].apply(lambda row: grouped.iloc[row, 2])

resid_cols = ['index', 'Name', 'Market_Value', 'Predicted Values', 'Residuals', 'Season', 'Position']
residuals = residuals[resid_cols]

#model coefficients
coefficients = list(zip(X_val.columns, ridge_model.coef_))

coefficients.sort(key = lambda t: t[1], reverse = True)
coefficients

"""with sns.axes_style('white'):
    plot=residuals.plot(kind='scatter',
                  x='Predicted Values',y='Residuals',alpha=0.3,figsize=(10,6));"""

#### ElasticNet CV

elastic = ElasticNetCV(alphas = alphavec, cv=5)
elastic.fit(X_tr, y_train) 

print('ElasticNet alpha = {}'.format(ridge_model.alpha_))

elastic_val_set_pred = elastic.predict(X_val_sc)

print('ElasticNet  MAE = {}'.format(mae(y_val, elastic_val_set_pred)))
print('ElasticNet Validation R^2 = {}'.format(r2_score(y_val, elastic_val_set_pred)))

elastic_test_set_pred = elastic.predict(X_test_sc)

print('ElasticNet  MAE = {}'.format(mae(y_test, elastic_test_set_pred)))
print('ElasticNet Test R^2 = {}'.format(r2_score(y_test, elastic_test_set_pred)))

coefficients = list(zip(X_test.columns, elastic.coef_))

coefficients.sort(key = lambda t: t[1], reverse = True)
coefficients

elastic_resids = y_test - elastic_test_set_pred
elastic_residuals = pd.DataFrame({'Residuals':elastic_resids, 'Predicted Values': elastic_test_set_pred})

pd.set_option('display.float_format', '{:,.1f}'.format)

elastic_residuals.reset_index(inplace = True)
elastic_residuals['Name'] = elastic_residuals['index'].apply(lambda row: grouped.iloc[row, [0]])
elastic_residuals['Market_Value'] = elastic_residuals['index'].apply(lambda row: grouped.iloc[row, 3])
elastic_residuals['Season'] = elastic_residuals['index'].apply(lambda row: grouped.iloc[row, 1])
elastic_residuals['Position'] = elastic_residuals['index'].apply(lambda row: grouped.iloc[row, 2])

resid_cols = ['index', 'Name', 'Market_Value', 'Predicted Values', 'Residuals', 'Season', 'Position']
elastic_residuals = elastic_residuals[resid_cols]

elastic_residuals.sort_values(by = 'Residuals', ascending = True).head(20)

## Random Forest Regressor

regr = RandomForestRegressor(max_depth=15, random_state=0, n_estimators=100, max_features = 70)
regr.fit(X_train_RF, y_train_RF) 

rf_predictions = regr.predict(X_val_RF)

print('Random Forest Validation R^2 = {}'.format(regr.score(X_val_RF, y_val_RF)))


random_resids = y_val_RF - rf_predictions
random_residuals = pd.DataFrame({'Residuals':random_resids, 'Predicted Values': rf_predictions})

random_residuals.reset_index(inplace = True)
random_residuals['Name'] = random_residuals['index'].apply(lambda row: grouped.iloc[row, [0]])
random_residuals['Market_Value'] = random_residuals['index'].apply(lambda row: grouped.iloc[row, 3])
random_residuals['Season'] = random_residuals['index'].apply(lambda row: grouped.iloc[row, 1])
random_residuals['Position'] = random_residuals['index'].apply(lambda row: grouped.iloc[row, 2])

random_residuals = random_residuals[resid_cols]

"print(regr.feature_importances_)"

## Random Forest with Random Search Cross Validation

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 100, stop = 1000, num = 10)]
# Number of features to consider at every split
max_features = ['auto', 'sqrt']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(10, 110, num = 11)]
max_depth.append(None)
# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]
# Method of selecting samples for training each tree
bootstrap = [True, False]
# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}

# Use the random grid to search for best hyperparameters
# First create the base model to tune
rf = RandomForestRegressor()
# Random search of parameters, using 3 fold cross validation, 
# search across 100 different combinations, and use all available cores
rf_random = RandomizedSearchCV(estimator = rf, param_distributions = random_grid, n_iter = 100, cv = 3, verbose=2, random_state=42, n_jobs = -1)
# Fit the random search model
rf_random.fit(X_train_RF, y_train_RF)

rf_random.best_params_

best_random = rf_random.best_estimator_

print('Best Random Forest Validation R^2 = {}'.format(best_random.score(X_val_RF, y_val_RF)))

best_random_predictions = best_random.predict(X_val_RF)
best_random_resids = y_val_RF - best_random_predictions

best_random_residuals = pd.DataFrame({'Residuals':best_random_resids, 'Predicted Values': best_random_predictions})
#with sns.axes_style('white'):
#    plot=residuals.plot(kind='scatter',
                 # x='Predicted Values',y='Residuals',alpha=0.3,figsize=(10,6));

best_random_residuals.reset_index(inplace = True)
best_random_residuals['Name'] = best_random_residuals['index'].apply(lambda row: grouped.iloc[row, [0]])
best_random_residuals['Market_Value'] = best_random_residuals['index'].apply(lambda row: grouped.iloc[row, 3])
best_random_residuals['Season'] = best_random_residuals['index'].apply(lambda row: grouped.iloc[row, 1])
best_random_residuals['Position'] = best_random_residuals['index'].apply(lambda row: grouped.iloc[row, 2])

resid_cols = ['index', 'Name', 'Market_Value', 'Predicted Values', 'Residuals', 'Season', 'Position']
best_random_residuals = best_random_residuals[resid_cols]

best_random_residuals[best_random_residuals['Season'] == 2016].sort_values(by = 'Residuals', ascending = True).head(4)


best_random_residuals[best_random_residuals['Name'] == 'Frenkie de Jong']

best_random_residuals.sort_values(by = 'Residuals', ascending = True).head(20)

best_random_residuals.sort_values(by = 'Residuals', ascending = False).head(10)

importances = best_random.feature_importances_
std = np.std([tree.feature_importances_ for tree in best_random.estimators_],
             axis=0)
indices = np.argsort(importances)[::-1]

for f in range(X_val_RF.shape[1]):
    print("%d. feature %d (%f)" % (f + 1, indices[f], importances[indices[f]]) + X_val_RF.columns[f] )
